+++ clients/go/ahp/reducers.go
@@ -1288,6 +1288,11 @@ func ApplyActionToChangeset(state *ahptypes.ChangesetState, action ahptypes.Stat
 		}
 		return ReduceOutcomeNoOp
 
+	case *ahptypes.ChangesetContentChangedAction:
+		state.Files = a.Files
+		state.Operations = a.Operations
+		return ReduceOutcomeApplied
+
 	case *ahptypes.ChangesetOperationsChangedAction:
 		state.Operations = a.Operations
 		return ReduceOutcomeApplied